Fix memory overflow in first load

+++ Source Code/QloudSync/Mac/Synchrony/Synchronizer.cs
@@ -51,7 +51,7 @@ namespace GreenQloud.Synchrony
         }
         
 
-        public int BytesTransferred {
+        public long BytesTransferred {
             get{
                 return remoteRepo.Connection.TransferSize;
             }
